renamed presenter/films-filter -> presenter/main-navigation, + вью и рефакт
